# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ID Creator Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type GeneratedIdStatus = typeof GeneratedIdStatus[keyof typeof GeneratedIdStatus];


export const GeneratedIdStatus = {
  success: 'success',
  failed: 'failed',
  pending: 'pending',
} as const;

export interface GeneratedId {
  id: number;
  email: string;
  username: string;
  password: string;
  /** @nullable */
  recoveryEmail?: string | null;
  status: GeneratedIdStatus;
  /** @nullable */
  errorMessage?: string | null;
  /** @nullable */
  batchId?: number | null;
  createdAt: string;
}

export interface IdListResponse {
  data: GeneratedId[];
  total: number;
  page: number;
  limit: number;
}

export interface IdStats {
  total: number;
  todayTotal: number;
  successCount: number;
  failedCount: number;
  pendingCount: number;
  successRate: number;
  todaySuccess: number;
  todayFailed: number;
}

export interface DeleteResult {
  success: boolean;
}

export interface ExportResult {
  data: string;
  count: number;
  format: string;
}

export type BatchStatus = typeof BatchStatus[keyof typeof BatchStatus];


export const BatchStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Batch {
  id: number;
  name: string;
  targetCount: number;
  status: BatchStatus;
  createdAt: string;
  /** @nullable */
  completedAt?: string | null;
  successCount: number;
  failCount: number;
  progress?: number;
}

export interface BatchInput {
  /**
     * @minimum 1
     * @maximum 10000
     */
  targetCount: number;
  name?: string;
}

export interface Settings {
  id: number;
  dailyTarget: number;
  scheduleTime: string;
  proxyEnabled: boolean;
  retryEnabled: boolean;
  retryMax: number;
  isActive: boolean;
  emailDomain: string;
  usernamePrefix: string;
}

export interface SettingsInput {
  /**
     * @minimum 1
     * @maximum 10000
     */
  dailyTarget?: number;
  scheduleTime?: string;
  proxyEnabled?: boolean;
  retryEnabled?: boolean;
  /**
     * @minimum 1
     * @maximum 10
     */
  retryMax?: number;
  isActive?: boolean;
  emailDomain?: string;
  usernamePrefix?: string;
}

export type ProxyStatus = typeof ProxyStatus[keyof typeof ProxyStatus];


export const ProxyStatus = {
  active: 'active',
  inactive: 'inactive',
  failed: 'failed',
} as const;

export interface Proxy {
  id: number;
  host: string;
  port: number;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  password?: string | null;
  isActive: boolean;
  status: ProxyStatus;
  /** @nullable */
  lastUsed?: string | null;
  createdAt: string;
}

export interface ProxyInput {
  host: string;
  port: number;
  username?: string;
  password?: string;
}

export interface ProxyToggle {
  isActive: boolean;
}

export type ListIdsParams = {
status?: ListIdsStatus;
page?: number;
limit?: number;
};

export type ListIdsStatus = typeof ListIdsStatus[keyof typeof ListIdsStatus];


export const ListIdsStatus = {
  success: 'success',
  failed: 'failed',
  pending: 'pending',
} as const;

export type ExportIdsParams = {
status?: ExportIdsStatus;
format?: ExportIdsFormat;
};

export type ExportIdsStatus = typeof ExportIdsStatus[keyof typeof ExportIdsStatus];


export const ExportIdsStatus = {
  success: 'success',
  failed: 'failed',
  all: 'all',
} as const;

export type ExportIdsFormat = typeof ExportIdsFormat[keyof typeof ExportIdsFormat];


export const ExportIdsFormat = {
  txt: 'txt',
  csv: 'csv',
} as const;