# lib/api-zod/src/generated/types/batchInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ID Creator Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface BatchInput {
  /**
     * @minimum 1
     * @maximum 10000
     */
  targetCount: number;
  name?: string;
}