# lib/api-zod/src/generated/types/idListResponse.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ID Creator Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { GeneratedId } from './generatedId';

export interface IdListResponse {
  data: GeneratedId[];
  total: number;
  page: number;
  limit: number;
}