# lib/api-zod/src/generated/types/idStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ID Creator Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface IdStats {
  total: number;
  todayTotal: number;
  successCount: number;
  failedCount: number;
  pendingCount: number;
  successRate: number;
  todaySuccess: number;
  todayFailed: number;
}