# lib/api-zod/src/generated/types/proxyInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ID Creator Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface ProxyInput {
  host: string;
  port: number;
  username?: string;
  password?: string;
}