# lib/api-zod/src/generated/types/settings.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ID Creator Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface Settings {
  id: number;
  dailyTarget: number;
  scheduleTime: string;
  proxyEnabled: boolean;
  retryEnabled: boolean;
  retryMax: number;
  isActive: boolean;
  emailDomain: string;
  usernamePrefix: string;
}